check api response for error
There was silent errors returned by the API since the status code is 200
even for errors, this checks this
