+++ src/clients/mod.rs
@@ -1,6 +1,8 @@
 pub mod api_client;
+mod api_error;
 mod api_structures;
 pub mod sparql_client;
 
 pub use api_client::{ApiClient, ObjectType, PropertyDataType};
+pub use api_error::ApiError;
 pub use sparql_client::SparqlClient;
